# static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ES7-WebServices",
        "version": "1.0.1",
        "title": "ES7 WebService"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1.0/Contenido": {
            "get": {
                "tags": ["Contenido de Escuela Sabatica"],
                "summary": "Contenido de una Escuela Sabatica",
                "description": "",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "es7",
                    "in": "query",
                    "description": "Escuela Sabatica",
                    "required": true,
                    "type": "integer"
                },
                {
                    "name": "tipo",
                    "in": "query",
                    "description": "Tipo de Contenido",
                    "required": true,
                    "type": "integer"
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/Contenido/lessons": {
            "get": {
                "tags": ["Contenido de Escuela Sabatica"],
                "summary": "Contenido de una Escuela Sabatica por Leccion",
                "description": "",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "es7",
                    "in": "query",
                    "description": "Escuela Sabatica",
                    "required": true,
                    "type": "integer"
                },
                {
                    "name": "lesson",
                    "in": "query",
                    "description": "Leccion",
                    "required": true,
                    "type": "integer"
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/Contenido/egw": {
            "get": {
                "tags": ["Contenido de Escuela Sabatica"],
                "summary": "Contenido Egw de una Escuela Sabatica por Leccion",
                "description": "",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "es7",
                    "in": "query",
                    "description": "Escuela Sabatica",
                    "required": true,
                    "type": "integer"
                },
                {
                    "name": "lesson",
                    "in": "query",
                    "description": "Leccion",
                    "required": true,
                    "type": "integer"
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/Contenido/{id}/": {
            "get": {
                "tags": ["Contenido de Escuela Sabatica"],
                "summary": "Encontrar un contenido por su Identificador",
                "produces": ["application/json"],
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "Identificador",
                    "required": true,
                    "type": "integer"
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "404": {
                        "description": "Contenido no encontrado"
                    }
                }
            }
        },
        "/v1.0/Contenido/create": {
            "post": {
                "tags": ["Contenido de Escuela Sabatica"],
                "summary": "Crear Contenido de Escuela Sabatica",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "Contenido de Escuela Sabatica",
                    "in": "body",
                    "description": "Contenido de Escuela Sabatica a crear",
                    "required": true,
                    "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "idEs": {
                                    "type": "integer"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "verse": {
                                    "type": "string"
                                },
                                "remoteColor": {
                                    "type": "string"
                                },
                                "egw": {
                                    "type": "integer"
                                },
                                "date": {
                                    "type": "string"
                                },
                                "versionCode": {
                                    "type": "integer"
                                },
                                "lessonNumber": {
                                    "type": "integer"
                                }
                            }
                        
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/Contenido/inactive": {
            "patch": {
                "tags": ["Contenido de Escuela Sabatica"],
                "summary": "Deshabilitar un Contenido de Escuela Sabatica",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "Contenido de Escuela Sabatica",
                    "in": "body",
                    "description": "Contenido de Escuela Sabatica a Deshabilitar",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
        "/v1.0/Contenido/update": {
            "put": {
                "tags": ["Contenido de Escuela Sabatica"],
                "summary": "Editar un Contenido de Escuela Sabatica",
                "description": "",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "name": "Contenido de Escuela Sabatica",
                    "in": "body",
                    "description": "Contenido de Escuela Sabatica a editar",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "idEs": {
                                "type": "integer"
                            },
                            "title": {
                                "type": "string"
                            },
                            "verse": {
                                "type": "string"
                            },
                            "remoteColor": {
                                "type": "string"
                            },
                            "egw": {
                                "type": "integer"
                            },
                            "date": {
                                "type": "string"
                            },
                            "versionCode": {
                                "type": "integer"
                            },
                            "lessonNumber": {
                                "type": "integer"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    }
                }
            }
        },
...
    }
}